Fix : 알람 관련 버그 픽스
