+++ presentation/src/main/java/com/stop/MainActivity.kt
@@ -1,9 +1,10 @@
 package com.stop
 
+import android.app.KeyguardManager
 import android.content.Context
+import android.content.Intent
 import android.graphics.Rect
 import android.os.Bundle
-import com.stop.databinding.ActivityMainBinding
 import android.view.MotionEvent
 import android.view.View
 import android.view.WindowManager
@@ -11,6 +12,8 @@ import android.view.inputmethod.InputMethodManager
 import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import androidx.navigation.fragment.NavHostFragment
+import com.stop.databinding.ActivityMainBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -19,11 +22,30 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         installSplashScreen()
+
         super.onCreate(savedInstanceState)
 
         binding = ActivityMainBinding.inflate(layoutInflater)
-        hideStatusBar()
         setContentView(binding.root)
+
+        // TODO 버튼을 통해서 소리나 진동 끄기로 바꿔야함 !!
+        val intent = Intent(this, SoundService::class.java)
+        this.stopService(intent)
+
+        setStartDestination()
+        hideStatusBar()
+        showOverLockScreen()
+    }
+
+    private fun setStartDestination() {
+        intent.extras?.getInt("ALARM_CODE")?.let {
+            val navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+            val inflater = navHostFragment.navController.navInflater
+            val graph = inflater.inflate(R.navigation.nav_graph)
+            graph.setStartDestination(R.id.alarmStartFragment)
+
+            navHostFragment.navController.graph = graph
+        }
     }
 
     private fun hideStatusBar() {
@@ -55,4 +77,20 @@ class MainActivity : AppCompatActivity() {
             this.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
         inputMethodManager.hideSoftInputFromWindow(view.windowToken, 0)
     }
+
+    private fun showOverLockScreen() {
+        window.addFlags(
+            WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON or
+                    WindowManager.LayoutParams.FLAG_DISMISS_KEYGUARD or
+                    WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED or
+                    WindowManager.LayoutParams.FLAG_TURN_SCREEN_ON
+        )
+
+        if (isMoreThanOreoMr1()) {
+            setShowWhenLocked(true)
+            setTurnScreenOn(true)
+            (getSystemService(Context.KEYGUARD_SERVICE) as KeyguardManager).requestDismissKeyguard(this, null)
+        }
+    }
+
 }
