--- presentation/src/main/java/com/stop/TestActivity.kt
@@ -1,36 +1,0 @@
-package com.stop
-
-import android.app.KeyguardManager
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import android.view.WindowManager
-import androidx.appcompat.app.AppCompatActivity
-
-class TestActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_test)
-
-        showOverLockScreen()
-
-        // TODO 버튼을 통해서 소리나 진동 끄기로 바꿔야함 !!
-        val intent = Intent(this, SoundService::class.java)
-        this.stopService(intent)
-    }
-
-    private fun showOverLockScreen() {
-        window.addFlags(
-            WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON or
-                    WindowManager.LayoutParams.FLAG_DISMISS_KEYGUARD or
-                    WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED or
-                    WindowManager.LayoutParams.FLAG_TURN_SCREEN_ON
-        )
-
-        if (isMoreThanOreoMr1()) {
-            setShowWhenLocked(true)
-            setTurnScreenOn(true)
-            (getSystemService(Context.KEYGUARD_SERVICE) as KeyguardManager).requestDismissKeyguard(this, null)
-        }
-    }
-}
